+++ src/Access/Categories.php
@@ -50,6 +50,6 @@ final class Categories {
 	 * @return int[]
 	 */
 	public function getCategoriesOfCategory($category_ref_id) {
-		return self::dic()->tree()->getSubTree(self::dic()->tree()->getNodeData($category_ref_id), false, self::TYPE_CAT);
+		return self::dic()->repositoryTree()->getSubTree(self::dic()->repositoryTree()->getNodeData($category_ref_id), false, self::TYPE_CAT);
 	}
 }